Remove old Stores.BTCLike controller
